requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
email-validator
aioimaplib
cryptography